Add SignalR support for message updates and deletions
Updated `MessagesController` to include `IHubContext<MessageHub>` for real-time notifications.
Enhanced `UpdateWithUserCheck` and `RemoveWithUserCheck` methods to fetch messages by ID, return `NotFound` if not found, and send SignalR messages upon successful updates and deletions.

+++ src/Messenger.API/Controllers/MessagesController.cs
@@ -4,13 +4,16 @@ using Messenger.Application.Mapper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Security.Claims;
+using Messenger.Application.Hubs;
+using Microsoft.AspNetCore.SignalR;
 
 namespace Messenger.API.Controllers
 {
     [ApiController]
     public class MessagesController(
         IMessageRepository messageRepository,
-        MessageMapper messageMapper
+        MessageMapper messageMapper,
+        IHubContext<MessageHub> hubContext
         ) : ControllerBase
     {
         [Authorize]
@@ -116,8 +119,13 @@ namespace Messenger.API.Controllers
             try
             {
                 var userId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
+                var message = await messageRepository.GetByIdAsync(dto.Id);
+                if (message is null)
+                    return NotFound("There is no message with this id.");
 
                 await messageRepository.UpdateWithUserCheckAsync(dto, userId);
+                
+                await hubContext.Clients.Group(message.ChatId.ToString()).SendAsync("OnMessageEdited", message.Id, dto.Content);
 
                 return NoContent();
             }
@@ -134,8 +142,13 @@ namespace Messenger.API.Controllers
             try
             {
                 var userId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
+                var message = await messageRepository.GetByIdAsync(id);
+                if (message is null)
+                    return NotFound("There is no message with this id.");
 
                 await messageRepository.DeleteWithUserCheckAsync(id, userId);
+
+                await hubContext.Clients.Group(message.ChatId.ToString()).SendAsync("OnMessageDeleted", message.ChatId, message.Id);
 
                 return NoContent();
             }
